+++ src/xo-app/backup/overview/index.js
@@ -10,7 +10,8 @@ import React from 'react'
 import SortedTable from 'sorted-table'
 import StateButton from 'state-button'
 import Tooltip from 'tooltip'
-import { addSubscriptions, constructQueryString } from 'utils'
+import { addSubscriptions } from 'utils'
+import { constructQueryString } from 'smart-backup-pattern'
 import { createSelector } from 'selectors'
 import { Card, CardHeader, CardBlock } from 'card'
 import { filter, find, forEach, get, map, orderBy } from 'lodash'
